MODQM-296 - refactored as requested
